Send webhook on player leave and fix channel ID parsing
Added asynchronous Discord webhook execution when a player leaves the server in ChatListener. Fixed channel ID parsing in Bot.java by switching from Integer to Long to support larger Discord channel IDs. Updated build number to 608.

+++ src/main/java/valorless/discordchat/discord/Bot.java
@@ -183,7 +183,7 @@ public class Bot implements Listener {
 		if(channel == null) {
 			try {
 				for(String ch : Bot.config.GetStringList("channels")) {
-					int id = Integer.valueOf(ch);
+					Long id = Long.valueOf(ch);
 					Guild guild = Main.bot.client.getGuildChannelById(id).getGuild();
 					GuildChannel gchannel = guild.getGuildChannelById(id);
 					if(guild.getSelfMember().hasPermission(gchannel, Permission.MESSAGE_SEND)) {
